+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+### Fixed
+- Fix bug in repr of Model class ([PR #35](https://github.com/alegonz/baikal/pull/35)).
 
 ## [0.4.0] - 2020-05-16
 ### Added